fix: update to new Zoe/spawner API
The agoric-sdk repo has been updated to use new-SES (see
https://github.com/Agoric/agoric-sdk/pull/1201 for details).

This changes the Zoe and spawner APIs: `zoe~.install(bundle)` instead of
`zoe~.install(source, moduleFormat)`, and `zoe.getInstallation` returns the
bundle object instead of a single string (so you must use a deep-equality
predicate).

The APIs are backwards-compatible, for now, but eventually support for the
old two-argument install() calls will be removed.